feat: updates logic for functional requirements and data persistence
